gc: compute reference map while computing the referrers closure, use in markAlive
This introduces a reference map which provides backlinks within the
referrers closure of a path. It uses this map in place of a call to
`computeFSClosure` in `markAlive`.

Because our reference map includes only paths that we've visited, we
will mark fewer nodes as 'alive'. This is fine -- any references that
we don't see will be hit on some other call to `deleteReferrersClosure`.

The resulting code will be faster since we avoid hitting the database or
redoing work to build the reference map (we just invert the referrers
map which we're already creating). It will also be faster because it
does less total work per top-level path, meaning that we better interleave
path deletion (which hits the disk and will block on low-level caching
if we do this in too-large batches) and computation (leaving time for
those low-level caches to be flushed).

More importantly, this opens the door in the next commit to us using the
same map to delete paths, which will improve interleaving even more.