gc: do depth-first search through referrers closure rather than breadth-fist
In the next commit we are going to detect cycles as we iterate. This is
pretty hard to do with a breadth-first search but pretty easy with a
depth-first search.

The GC algorithm is ambivalent to which one we do, so switch.

We use a std::vector for the stack. There is also a std::stack but it
has an uniquely named API (using `top` as the accessor and `pop` for
removal). I needed to add a generic `pop_back` method to libutil/util.hh
and stuck with vector because it has more universal back()/pop_back()
names.