gc: move referrersCache into deleteReferrersClosure
Right now we have a cache of referrers which is maintained across calls
of deleteReferrersClosure, because we may encounter the same paths in
the referrers closures of multiple paths.

In the next commits we will rearrange the algorithm so that this doesn't
happen. Instead, when we compute the referrers closure of a path, we
will iterate through the entire closure and mark every single path as
either "alive" or "dead" (by copying it into the appropriate map).

This means that entries of the referrersCache will never be reused, and
it is simply a waste of memory to maintain a cache of all paths like
this. Furthermore, by moving this into deleteReferrersClosure, we can
repurpose it as a "visited" set, eliminating the `visited` map. (The
doccomment on the variable explains that it represents "visited but
unprocessed" paths, and that entries will be deleted when paths are
marked as alive or dead. This isn't true yet in this commit, but it will
be.)

This commit is likely to be a slowdown since we haven't yet changed the
algorithm to eliminate early returns and always process the entire
closure. But by the end of this PR we will correct that. Meanwhile,
this commit *is* a significant memory reduction.