gc: rather than deleting all non-alive nodes at the end, delete them inline
This is a fairly significant change to the GC algorithm.

Prior to this change, the way the GC worked was:

1. For each path in the store, compute its referrers closure by
   breadth-first search, skipping any previously-seen paths to avoid
   looping in case of cycles. Maintain a set `visited` of the closure
   as it's been computed so far.
2. If any path encountered is marked dead then skip it.
3. If any path encountered is marked alive (typically, a gc root or a
   descendant), then mark that path's reference closure (computed by
   database lookup) alive and bail out.
4. If this *doesn't* happen, then we know that the entire referrers
   closure is dead, so delete the whole thing in topological order.

This has two problems: one is that it's fairly CPU- and
memory-intensive, requiring multiple sets and maps. Because it might
consider paths multiple times (every time it sees an 'alive' path it
bails out of the current referrers closure, potentially wasting work)
it maintains a global referrersCache object which avoids pinging the
database all the time. (A previous commit made this non-global which
reduces memory at the expense of repeating even -more- work.) It stores
the referrers closure as a set `visited` and then creates a whole copy
of it by calling topoSortPaths. (And again, a previous commit made this
even worse by replacing the set with the key-set of `referrersCache`.)

The second problem is that it's bursty: it does all this computation to
produce a giant blob of things to delete and then deletes them all at
once. This quickly saturates disk caches and especially for spinning
disks results in a pattern where we're CPU-bound, then IO-bound, then
CPU-bound, and so on.

INSTEAD, this commit changes the algorithm to work as follows:

1. For each path in the store, compute its referrers closure by
   depth-first search, maintain a map `referrersCache` of the referrers
   relations seen so far as well as the inverse map `referenceCache`.

   Treat `referenceCache` as we did `visited`; this represents the
   "unprocessed referrers closure as it has been computed so far".
2. If we see a known-dead path then just don't follow it, as before.
3. If we see a known-alive path, then follow `referenceCache` backward
   to mark the reference closure (well, the subset of the reference
   closure that we've seen so far) alive; remove all these alive paths
   from `referenceCache` since they're no longer "unprocessed" and
   continue.
4. When we encounter a cycle (we maintain a trace of our depth-first
   search so can check for these easily) we *don't* record the cyclic
   reference in our `referrersCache`, because this object is used to
   distingush roots from non-roots and cyclic references don't turn
   roots into non-roots) (conceptually, they turn a root into a "big
   root" consisting of an un-referred-to cycle). We *do* record the
   cyclic reference in our `referenceCache`, since this object is used
   to propagate liveness from gcroots up to all its descendants.
5. When we encounter a root (other than a gcroot or temporary root,
   which is handled in step 3), we delete it, both from the store
   (doing whatever disk access that entails), delete it from
   `referenceCache` (since it's no longer unprocessed), and delete
   it from the `referrersCache` entries of all its references. We then
   recurse on any of the references which have become roots.

This algorithm, for each path in the store, computes its entire
referrers closure. For every path in this closure, we mark it alive or
dead. If it's marked dead, we delete it from the store at the same time,
interleaving disk accesses with graph-traversal work much better and
(hopefully) doing useful work while IO caches are being flushed behind
the scenes.

Because it processes the entire closure we don't need to maintain any
global tables, except for the alive/dead sets, saving memory by
eliminating the old `referrersCache` map.